Split the views module into class per file
