+++ src/annotation/urls.py
@@ -1,11 +1,11 @@
-from django.urls import path
-
+"""Defines the routing table of the application."""
 from . import views
+from django.urls import path
 
 app_name = "annotation"
 urlpatterns = [
     path("", views.IndexView.as_view(), name="index"),
-    path("new", views.NewAnnotatioView.as_view(), name="new-annotation"),
+    path("new", views.NewAnnotationView.as_view(), name="new-annotation"),
     path("<int:id>", views.AnnotateView.as_view(), name="annotate"),
     path("api/entries/<int:entry_id>",
          views.GetEntryContentsView.as_view(),
@@ -14,5 +14,5 @@ urlpatterns = [
          views.MarkAnnotationCompleteView.as_view(),
          name="mark-complete"),
     path("save", views.SaveAnnotationView.as_view(), name="save"),
-    path("thank-you", views.thank_you, name="thank-you"),
+    path("thank-you", views.ThankYouView.as_view(), name="thank-you"),
 ]
