+++ src/annotation/views/__init__.py
@@ -1,0 +1,8 @@
+"""Defines the views of the application."""
+from .annotate import AnnotateView
+from .getentrycontents import GetEntryContentsView
+from .index import IndexView
+from .markcomplete import MarkAnnotationCompleteView
+from .newannotation import NewAnnotationView
+from .saveannotation import SaveAnnotationView
+from .thankyou import ThankYouView
