+++ src/annotation/views/viewsettings.py
@@ -1,0 +1,5 @@
+"""Defines the view settings."""
+from django.conf import settings
+
+MAX_CONCURRENT_ANNOTATORS = int(
+    getattr(settings, "MAX_CONCURRENT_ANNOTATORS", 2))
